Collapse: migrate modal.svelte in place, drop parallel runes file
