+++ src/lib/components/data-encoder-settings.svelte
@@ -10,7 +10,7 @@
   import { page } from '$app/state';
 
   import Link from '$lib/holocene/link.svelte';
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import RadioGroup from '$lib/holocene/radio-input/radio-group.svelte';
   import RadioInput from '$lib/holocene/radio-input/radio-input.svelte';
   import { translate } from '$lib/i18n/translate';
