+++ src/lib/components/schedule/schedule-action-modals/delete-schedule-modal.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import { translate } from '$lib/i18n/translate';
   import {
     actionPending,
