+++ src/lib/components/schedule/schedule-form/cron-expression-format-modal.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import { translate } from '$lib/i18n/translate';
 
   interface Props {
