+++ src/lib/components/standalone-activities/cancel-confirmation-modal.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import { translate } from '$lib/i18n/translate';
   import { cancelActivityExecution } from '$lib/services/standalone-activities';
   import { toaster } from '$lib/stores/toaster';
