+++ src/lib/components/workflow/client-actions/batch-terminate-confirmation-modal.svelte
@@ -4,7 +4,7 @@
   import { getContext } from 'svelte';
 
   import BatchOperationConfirmationModalBody from '$lib/components/batch/batch-operation-confirmation-form.svelte';
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import { translate } from '$lib/i18n/translate';
   import Translate from '$lib/i18n/translate.svelte';
   import { Action } from '$lib/models/workflow-actions';
