+++ src/lib/components/workflow/client-actions/terminate-confirmation-modal.svelte
@@ -2,7 +2,7 @@
   import type { Writable } from 'svelte/store';
 
   import Input from '$lib/holocene/input/input.svelte';
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import { translate } from '$lib/i18n/translate';
   import { Action } from '$lib/models/workflow-actions';
   import { terminateWorkflow } from '$lib/services/workflow-service';
