+++ src/lib/holocene/modal.stories.svelte
@@ -2,6 +2,7 @@
 
 <script lang="ts" module>
   import type { Meta } from '@storybook/svelte';
+  import type { ComponentProps } from 'svelte';
 
   import Modal from './modal.svelte';
 
@@ -84,10 +85,8 @@
         name: 'Highlight Navigation',
         control: 'boolean',
       },
-
-      toggleModal: { table: { disable: true } },
     },
-  } satisfies Meta<Modal>;
+  } satisfies Meta<ComponentProps<typeof Modal>>;
 </script>
 
 <script lang="ts">
@@ -98,25 +97,30 @@
 <Template let:args>
   <Modal
     id="modal"
-    on:confirmModal={action('confirm')}
-    on:cancelModal={action('cancel')}
+    onConfirmModal={action('confirm')}
+    onCancelModal={action('cancel')}
     {...args}
   >
-    <h3 slot="title">Modal Title</h3>
-    <p slot="content">
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus.
-      Suspendisse lectus tortor, dignissim sit amet, adipiscing nec, ultricies
-      sed, dolor. Cras elementum ultrices diam. Maecenas ligula massa, varius a,
-      semper congue, euismod non, mi. Proin porttitor, orci nec nonummy
-      molestie, enim est eleifend mi, non fermentum diam nisl sit amet erat.
-      Duis semper. Duis arcu massa, scelerisque vitae, consequat in, pretium a,
-      enim. Pellentesque congue. Ut in risus volutpat libero pharetra tempor.
-      Cras vestibulum bibendum augue. Praesent egestas leo in pede. Praesent
-      blandit odio eu enim. Pellentesque sed dui ut augue blandit sodales.
-      Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere
-      cubilia Curae; Aliquam nibh. Mauris ac mauris sed pede pellentesque
-      fermentum. Maecenas adipiscing ante non diam sodales hendrerit.
-    </p>
+    {#snippet titleSnippet()}
+      <h3>Modal Title</h3>
+    {/snippet}
+    {#snippet content()}
+      <p>
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus.
+        Suspendisse lectus tortor, dignissim sit amet, adipiscing nec, ultricies
+        sed, dolor. Cras elementum ultrices diam. Maecenas ligula massa, varius
+        a, semper congue, euismod non, mi. Proin porttitor, orci nec nonummy
+        molestie, enim est eleifend mi, non fermentum diam nisl sit amet erat.
+        Duis semper. Duis arcu massa, scelerisque vitae, consequat in, pretium
+        a, enim. Pellentesque congue. Ut in risus volutpat libero pharetra
+        tempor. Cras vestibulum bibendum augue. Praesent egestas leo in pede.
+        Praesent blandit odio eu enim. Pellentesque sed dui ut augue blandit
+        sodales. Vestibulum ante ipsum primis in faucibus orci luctus et
+        ultrices posuere cubilia Curae; Aliquam nibh. Mauris ac mauris sed pede
+        pellentesque fermentum. Maecenas adipiscing ante non diam sodales
+        hendrerit.
+      </p>
+    {/snippet}
   </Modal>
 </Template>
