+++ src/lib/holocene/modal.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import type { HTMLAttributes } from 'svelte/elements';
 
-  import { type ComponentProps, createEventDispatcher } from 'svelte';
+  import type { ComponentProps, Snippet } from 'svelte';
   import { twMerge as merge } from 'tailwind-merge';
 
   import Button from '$lib/holocene/button.svelte';
 
   import IconButton from './icon-button.svelte';
 
-  interface $$Props extends HTMLAttributes<HTMLDialogElement> {
+  interface Props extends HTMLAttributes<HTMLDialogElement> {
     cancelText: string;
     confirmDisabled?: boolean;
     confirmText: string;
@@ -19,33 +19,41 @@
     id: string;
     large?: boolean;
     loading?: boolean;
-    'data-testid'?: string;
     open: boolean;
     error?: string;
     class?: string;
+    titleSnippet?: Snippet;
+    content?: Snippet;
+    footer?: Snippet;
+    onCancelModal?: () => void;
+    onConfirmModal?: () => void;
   }
 
-  export let hideCancel = false;
-  export let hideConfirm = false;
-  export let confirmText: string;
-  export let cancelText: string;
-  export let confirmType: ComponentProps<typeof Button>['variant'] = 'primary';
-  export let confirmDisabled = false;
-  export let large = false;
-  export let loading = false;
-  export let hightlightNav = false;
-  export let id: string;
-  export let open: boolean;
-  export let error = '';
+  let {
+    hideCancel = false,
+    hideConfirm = false,
+    confirmText,
+    cancelText,
+    confirmType = 'primary',
+    confirmDisabled = false,
+    large = false,
+    loading = false,
+    hightlightNav = false,
+    id,
+    open = $bindable(),
+    error = $bindable(''),
+    class: className = '',
+    titleSnippet,
+    content,
+    footer,
+    onCancelModal,
+    onConfirmModal,
+    ...rest
+  }: Props = $props();
 
-  let className = '';
-  export { className as class };
+  let modalElement = $state<HTMLDialogElement>();
 
-  let modalElement: HTMLDialogElement;
-
-  $: toggleModal(open, modalElement);
-
-  export const toggleModal = (open: boolean, modal: HTMLDialogElement) => {
+  const toggleModal = (open: boolean, modal: HTMLDialogElement | undefined) => {
     if (open) {
       modal?.showModal();
     } else {
@@ -53,19 +61,18 @@
     }
   };
 
-  const dispatch = createEventDispatcher<{
-    cancelModal: undefined;
-    confirmModal: undefined;
-  }>();
+  $effect(() => {
+    toggleModal(open, modalElement);
+  });
 
   const handleCancel = () => {
-    dispatch('cancelModal');
+    onCancelModal?.();
     open = false;
     error = '';
   };
 
   const confirmModal = () => {
-    dispatch('confirmModal');
+    onConfirmModal?.();
   };
 
   const closeModal = () => {
@@ -76,18 +83,18 @@
     if (event.target === modalElement) closeModal();
   };
 
-  $: {
+  $effect(() => {
     if (open && modalElement) {
       modalElement.focus();
     }
-  }
+  });
 </script>
 
-<svelte:window on:click={handleClick} />
+<svelte:window onclick={handleClick} />
 
 <dialog
   {id}
-  on:close={handleCancel}
+  onclose={handleCancel}
   bind:this={modalElement}
   class={merge(
     'body',
@@ -100,8 +107,7 @@
   class:hightlightNav
   aria-modal="true"
   aria-labelledby="modal-title-{id}"
-  data-testid={$$props['data-testid']}
-  {...$$restProps}
+  {...rest}
 >
   {#if !loading}
     <IconButton
@@ -112,11 +118,17 @@
     />
   {/if}
   <div id="modal-title-{id}" class="title">
-    <slot name="title" />
+    {@render titleSnippet?.()}
   </div>
-  <form on:submit|preventDefault={confirmModal} method="dialog">
+  <form
+    onsubmit={(event) => {
+      event.preventDefault();
+      confirmModal();
+    }}
+    method="dialog"
+  >
     <div id="modal-content-{id}" class="content">
-      <slot name="content" />
+      {@render content?.()}
       <p
         class="mt-2 text-sm font-normal text-danger"
         class:hidden={!error}
@@ -127,9 +139,11 @@
     </div>
 
     <div class="flex items-center justify-between p-6">
-      <slot name="footer">
+      {#if footer}
+        {@render footer()}
+      {:else}
         <div></div>
-      </slot>
+      {/if}
       <div class="flex items-center justify-end space-x-2">
         {#if !hideCancel}
           <Button
