+++ src/lib/pages/nexus-edit-endpoint.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import Button from '$lib/holocene/button.svelte';
   import Input from '$lib/holocene/input/input.svelte';
-  import Modal from '$lib/holocene/modal-runes.svelte';
+  import Modal from '$lib/holocene/modal.svelte';
   import { translate } from '$lib/i18n/translate';
   import NexusForm, {
     type NexusFormData,
